server: add WithGrpcServer
Add WithGrpcServer option for the New function.

WithGrpcServer sets the gRPC server that the Server will be
started on. It also calls Stop for the previously-configured
grpcServer, which ensures that channelz entries are cleaned
up (channelz entries are created at the point of calling
grpc.NewServer).

This change also adds checks at (*Server).Serve and
(*Server).Halt to ensure that they are only called for
properly instantiated *Server structs.